Add auth

+++ app/Http/Controllers/ThreadsController.php
@@ -84,11 +84,10 @@ class ThreadsController extends Controller
         ]);
     }
 
-    /**
-     * 
-     */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         if(request()->wantsJson()) {
